Remove old implementation of Crashes
