+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/capture/crash/CrashDataSourceImplTest.kt
@@ -3,6 +3,9 @@ package io.embrace.android.embracesdk.capture.crash
 import io.embrace.android.embracesdk.FakeNdkService
 import io.embrace.android.embracesdk.arch.assertIsType
 import io.embrace.android.embracesdk.arch.schema.EmbType
+import io.embrace.android.embracesdk.config.local.CrashHandlerLocalConfig
+import io.embrace.android.embracesdk.config.local.LocalConfig
+import io.embrace.android.embracesdk.config.local.SdkLocalConfig
 import io.embrace.android.embracesdk.fakes.FakeAnrService
 import io.embrace.android.embracesdk.fakes.FakeConfigService
 import io.embrace.android.embracesdk.fakes.FakeCrashFileMarker
@@ -10,6 +13,7 @@ import io.embrace.android.embracesdk.fakes.FakeLogOrchestrator
 import io.embrace.android.embracesdk.fakes.FakeLogWriter
 import io.embrace.android.embracesdk.fakes.FakePreferenceService
 import io.embrace.android.embracesdk.fakes.FakeSessionOrchestrator
+import io.embrace.android.embracesdk.fakes.fakeAutoDataCaptureBehavior
 import io.embrace.android.embracesdk.fakes.fakeEmbraceSessionProperties
 import io.embrace.android.embracesdk.internal.serialization.EmbraceSerializer
 import io.embrace.android.embracesdk.logging.EmbLogger
@@ -62,7 +66,18 @@ internal class CrashDataSourceImplTest {
         testException = RuntimeException("Test exception")
     }
 
-    private fun setupForHandleCrash() {
+    private fun setupForHandleCrash(crashHandlerEnabled: Boolean = false) {
+        configService = FakeConfigService(
+            autoDataCaptureBehavior = fakeAutoDataCaptureBehavior(
+                localCfg = {
+                    LocalConfig(
+                        "",
+                        false,
+                        SdkLocalConfig(crashHandler = CrashHandlerLocalConfig(crashHandlerEnabled))
+                    )
+                }
+            )
+        )
         crashDataSource = CrashDataSourceImpl(
             logOrchestrator,
             sessionOrchestrator,
@@ -98,7 +113,10 @@ internal class CrashDataSourceImplTest {
         assertEquals(1, anrService.forceAnrTrackingStopOnCrashCount)
         assertEquals(1, logWriter.logEvents.size)
         val lastSentCrash = logWriter.logEvents.single()
-        assertEquals(logWriter.logEvents.single().schemaType.attributes()["log.record.uid"], sessionOrchestrator.crashId)
+        assertEquals(
+            logWriter.logEvents.single().schemaType.attributes()["log.record.uid"],
+            sessionOrchestrator.crashId
+        )
 
         /*
          * Verify mainCrashHandled is true after the first execution
@@ -119,7 +137,10 @@ internal class CrashDataSourceImplTest {
 
         assertEquals(1, anrService.forceAnrTrackingStopOnCrashCount)
         assertEquals(1, logWriter.logEvents.size)
-        assertEquals(logWriter.logEvents.single().schemaType.attributes()["log.record.uid"], sessionOrchestrator.crashId)
+        assertEquals(
+            logWriter.logEvents.single().schemaType.attributes()["log.record.uid"],
+            sessionOrchestrator.crashId
+        )
         assertEquals(ndkService.lastUnityCrashId, sessionOrchestrator.crashId)
     }
 
@@ -152,5 +173,17 @@ internal class CrashDataSourceImplTest {
                 "at com.example.MyClass.method(MyClass.kt:10)\"}",
             lastSentCrashAttributes["emb.android.react_native_crash.js_exception"]
         )
+    }
+
+    @Test
+    fun `test exception handler is registered with config option enabled`() {
+        setupForHandleCrash(true)
+        assert(Thread.getDefaultUncaughtExceptionHandler() is EmbraceUncaughtExceptionHandler)
+    }
+
+    @Test
+    fun `test exception handler is not registered with config option disabled`() {
+        setupForHandleCrash(false)
+        assert(Thread.getDefaultUncaughtExceptionHandler() !is EmbraceUncaughtExceptionHandler)
     }
 }
