+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/FakeModuleInitBootstrapper.kt
@@ -52,7 +52,7 @@ internal fun fakeModuleInitBootstrapper(
     nativeModuleSupplier: NativeModuleSupplier = { _, _, _, _, _, _, _ -> FakeNativeModule() },
     dataContainerModuleSupplier: DataContainerModuleSupplier = { _, _, _, _, _ -> FakeDataContainerModule() },
     sessionModuleSupplier: SessionModuleSupplier = { _, _, _, _, _, _, _, _, _, _, _, _ -> FakeSessionModule() },
-    crashModuleSupplier: CrashModuleSupplier = { _, _, _, _, _, _, _, _, _, _ -> FakeCrashModule() },
+    crashModuleSupplier: CrashModuleSupplier = { _, _, _, _, _, _, _, _ -> FakeCrashModule() },
     payloadModuleSupplier: PayloadModuleSupplier = { _, _, _, _, _, _, _, _, _, _, _ -> FakePayloadModule() }
 ) = ModuleInitBootstrapper(
     logger = fakeEmbLogger,
